+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - **BREAKING**: Add support for `project` and `path` input methods in config file
+- Add directives config option
 
 # 0.1.0